requirements.txt
pdfminer.six
pandas
numpy
matplotlib
yfinance
plotly
nltk
